jc/Restructured View-Form Feature folder to aid with implementation
