+++ frontend/src/components/Admin/Inventory/Forms/View-Form/Info-Tab/Info-Tab.tsx
@@ -1,17 +1,13 @@
 "use client";
 
 import dynamic from "next/dynamic";
-import Link from "next/link";
-import { useParams, useRouter } from "next/navigation";
-import { useState, useEffect } from "react";
-import { getItemById, updateItem, deleteItem, Item } from "@/api/items";
-import { getSalesByItemId, Sale } from "@/api/sales";
+import { useState } from "react";
+import { Item } from "@/api/items";
+import { Sale } from "@/api/sales";
 import EditableField from "@/components/ui/EditableField";
 import SelectField, { SelectOption } from "@/components/ui/selectField";
 import ItemImageUpload from "@/components/Admin/Inventory/Upload-Image/ItemImageUpload";
 import SetCoverPhotoModal from "@/components/Admin/Inventory/SetCoverPhotoModal";
-import SaleModal from "@/components/Admin/Inventory/Forms/Add-Sale";
-import { Trash, Printer } from "lucide-react";
 import { Card } from "@/components/ui/card";
 
 const GRADE_OPTIONS: SelectOption[] = [
@@ -68,219 +64,59 @@ function Field({
   );
 }
 
-export default function ViewForm() {
-  const [item, setItem] = useState<Item | null>(null);
-  const [sales, setSales] = useState<Sale[]>([]);
-  const [salesLoading, setSalesLoading] = useState(true);
-  const [images, setImages] = useState<string[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [formData, setFormData] = useState<Partial<Item>>({});
-  const [saving, setSaving] = useState(false);
+const formatCurrency = (value?: number) =>
+  new Intl.NumberFormat("en-US", {
+    style: "currency",
+    currency: "USD",
+  }).format(value ?? 0);
+
+const formatWeight = (value?: number, unit?: "kg" | "lb") => {
+  const weight = value ?? 0;
+  const suffix = unit === "kg" ? "kg" : "lbs";
+  return `${weight.toLocaleString("en-US", {
+    maximumFractionDigits: 1,
+    minimumFractionDigits: Number.isInteger(weight) ? 0 : 1,
+  })} ${suffix}`;
+};
+
+type InfoTabProps = {
+  itemId: string;
+  item: Item;
+  formData: Partial<Item>;
+  setFormData: React.Dispatch<React.SetStateAction<Partial<Item>>>;
+  images: string[];
+  setImages: React.Dispatch<React.SetStateAction<string[]>>;
+  sales: Sale[];
+  salesLoading: boolean;
+  saving: boolean;
+  onPublish: () => void;
+  onSave: () => void;
+};
+
+export default function InfoTab({
+  itemId,
+  item,
+  formData,
+  setFormData,
+  images,
+  setImages,
+  sales,
+  salesLoading,
+  saving,
+  onPublish,
+  onSave,
+}: InfoTabProps) {
   const [showCoverModal, setShowCoverModal] = useState(false);
-  const [showSaleModal, setShowSaleModal] = useState(false);
-  const [toast, setToast] = useState<{ message: string; sub: string } | null>(null);
 
-  const { id: itemId } = useParams<{ id: string }>();
-  const qrValue = formData.qrCode ?? item?.qrCode ?? itemId;
-  const router = useRouter();
-
-  useEffect(() => {
-    const fetchItem = async () => {
-      try {
-        const data = await getItemById(itemId);
-        setItem(data);
-        setFormData(data);
-        setImages(data.images ?? []);
-      } catch {
-        setError("Failed to load item.");
-      } finally {
-        setLoading(false);
-      }
-    };
-    fetchItem();
-  }, [itemId]);
-
-  useEffect(() => {
-    const fetchSales = async () => {
-      try {
-        setSalesLoading(true);
-        const data = await getSalesByItemId(itemId);
-        setSales(data);
-      } catch {
-        setSales([]);
-      } finally {
-        setSalesLoading(false);
-      }
-    };
-    fetchSales();
-  }, [itemId]);
+  const qrValue = formData.qrCode ?? item.qrCode ?? itemId;
 
   const set = (field: keyof Item) => (val: string) =>
     setFormData((prev) => ({ ...prev, [field]: val }));
 
-  const showToast = (message: string, sub: string) => {
-    setToast({ message, sub });
-    setTimeout(() => setToast(null), 6000);
-  };
-
-  const formatCurrency = (value?: number) =>
-    new Intl.NumberFormat("en-US", {
-      style: "currency",
-      currency: "USD",
-    }).format(value ?? 0);
-
-  const formatWeight = (value?: number, unit?: "kg" | "lb") => {
-    const weight = value ?? 0;
-    const suffix = unit === "kg" ? "kg" : "lbs";
-    return `${weight.toLocaleString("en-US", {
-      maximumFractionDigits: 1,
-      minimumFractionDigits: Number.isInteger(weight) ? 0 : 1,
-    })} ${suffix}`;
-  };
-
-  const handleGenerateLabel = () => {
-    const rows = [
-      ["SKU", "Breed", "Grade", "Farmer Name", "QR Code"],
-      [
-        formData.sku ?? "",
-        formData.breed ?? "",
-        formData.grade ?? "",
-        formData.farmerName ?? "",
-        formData.qrCode ?? "",
-      ],
-    ];
-
-    const csv = rows
-      .map((row) =>
-        row.map((cell) => `"${String(cell).replace(/"/g, '""')}"`).join(",")
-      )
-      .join("\n");
-
-    const blob = new Blob([csv], { type: "text/csv;charset=utf-8;" });
-    const url = URL.createObjectURL(blob);
-    const link = document.createElement("a");
-    link.href = url;
-    link.download = `label-${formData.sku ?? itemId}.csv`;
-    link.click();
-    URL.revokeObjectURL(url);
-  };
-
-  const handlePublish = async () => {
-    try {
-      if (formData?.isPublic == false) {
-        await updateItem(itemId, { isPublic: true });
-        setFormData((p) => ({ ...p, isPublic: true }));
-        showToast(
-          "Lot successfully published!",
-          "Your changes have been saved and published externally. Change the status to “Processing” to hide this lot from the public inventory.",
-        );
-      } else {
-        await updateItem(itemId, { isPublic: false });
-        setFormData((p) => ({ ...p, isPublic: false }));
-        showToast(
-          "Lot unpublished.",
-          "This lot is no longer visible in the public inventory.",
-        );
-      }
-    } catch {
-      setError("Failed to publish.");
-    }
-  };
-
-  const handleSave = async () => {
-    try {
-      setSaving(true);
-      await updateItem(itemId, {
-        ...formData,
-        images,
-        coverImage: images[0] ?? "",
-      });
-      showToast(
-        "Lot successfully updated!",
-        "Your changes have been saved internally. Change the status to “In Stock” to make this lot available for sale.",
-      );
-    } catch {
-      setError("Failed to save.");
-    } finally {
-      setSaving(false);
-    }
-  };
-
-  const handleDelete = async () => {
-    if (
-      !confirm("Unpublish this lot? Placing this lot on hold will remove this lot from the external inventory page.")
-    ) {
-      return;
-    }
-    try {
-      await deleteItem(itemId);
-      router.push("/inventory");
-    } catch {
-      setError("Failed to delete.");
-    }
-  };
-
-  if (loading) return <div className="p-4 md:p-8">Loading...</div>;
-  if (!item)
-    return (
-      <div className="p-4 md:p-8 text-red-600">{error ?? "Item not found"}</div>
-    );
-
   return (
-    <main className="min-h-screen bg-white px-4 py-6 md:p-8">
-      <div className="flex items-center justify-between mb-4 md:mb-8">
-        <Link
-          href="/inventory"
-          className="flex items-center gap-1 text-sm text-gray-600 hover:text-gray-900"
-        >
-          ← Back to Inventory
-        </Link>
-        <div className="hidden md:flex gap-2">
-          <button
-            onClick={handleGenerateLabel}
-            className="hover:bg-gray-50 rounded p-1"
-            aria-label="Print Label"
-          >
-            <Printer size={24} />
-          </button>
-          <button onClick={handleDelete}>
-            <Trash size={24} />
-          </button>
-          <button
-            onClick={() => setShowSaleModal(true)}
-            className="rounded bg-[#D9D9D9] px-4 py-1.5 text-sm text-black hover:bg-blue-500"
-          >
-            Record a sale
-          </button>
-          <button
-            onClick={handlePublish}
-            className="rounded bg-[#D9D9D9] px-4 py-1.5 text-sm text-black hover:bg-blue-500"
-          >
-            {formData.isPublic ? "Unpublish" : "Publish"}
-          </button>
-          <button
-            onClick={handleSave}
-            disabled={saving}
-            className="rounded bg-[#3A4F0D] px-4 py-1.5 text-sm text-white hover:bg-blue-500 disabled:opacity-50"
-          >
-            {saving ? "Saving..." : "Save"}
-          </button>
-        </div>
-      </div>
-
-      <div className="mb-6 hidden md:block">
-        <p className="text-base font-semibold text-gray-900">
-          SKU: {formData.sku ?? ""}
-        </p>
-      </div>
-
-      <h1 className="text-lg font-bold mb-4 md:hidden">
-        SKU: {formData.sku ?? itemId}
-      </h1>
-
-      <div className="grid grid-cols-1 md:grid-cols-[1fr_380px] gap-6 md:gap-12 items-start">
-        <div className="flex flex-col gap-6 md:gap-10">
+    <>
+      <div className="grid grid-cols-1 md:grid-cols-[minmax(0,1fr)_380px] gap-6 md:gap-12 items-start">
+        <div className="flex flex-col gap-6 md:gap-10 min-w-0">
           <div className="md:hidden">
             <p className="text-sm font-semibold text-gray-900 mb-2">Photos</p>
             <ItemImageUpload
@@ -302,7 +138,6 @@ export default function ViewForm() {
 
           <section>
             <Card className="p-6">
-
               <h2 className="text-lg font-bold mb-4 md:text-2xl md:mb-5">
                 General Information
               </h2>
@@ -339,7 +174,7 @@ export default function ViewForm() {
                   />
                 </Field>
               </div>
-            </Card> 
+            </Card>
           </section>
 
           <div className="md:hidden">
@@ -465,13 +300,13 @@ export default function ViewForm() {
           <div className="md:hidden pb-4">
             <div className="flex flex-col gap-2">
               <button
-                onClick={handlePublish}
+                onClick={onPublish}
                 className="w-full rounded-lg bg-[#9F9E97] px-4 py-3 text-sm text-white hover:bg-[#8a897e]"
               >
                 {formData.isPublic ? "Unpublish" : "Publish"}
               </button>
               <button
-                onClick={handleSave}
+                onClick={onSave}
                 disabled={saving}
                 className="w-full rounded-lg bg-[#2C2C2C] px-4 py-3 text-sm text-white hover:bg-[#1A1A1A] disabled:opacity-50"
               >
@@ -540,29 +375,6 @@ export default function ViewForm() {
           onSave={(newCover) => setFormData((p) => ({ ...p, coverImage: newCover }))}
         />
       )}
-
-      {showSaleModal && (
-        <SaleModal
-          itemId={itemId}
-          costPerWeight={formData.purchasePrice ?? 0}
-          onClose={() => setShowSaleModal(false)}
-          onSaleRecorded={(id, total) => console.log("Sold!", id, total)}
-        />
-      )}
-
-      {toast && (
-        <div className="fixed bottom-6 right-6 z-50 w-80 rounded-lg border border-gray-200 bg-white p-4 shadow-lg">
-          <div className="flex items-start justify-between gap-2">
-            <div>
-              <p className="font-semibold text-gray-900 text-sm">{toast.message}</p>
-              <p className="mt-1 text-sm text-gray-600">{toast.sub}</p>
-            </div>
-            <button onClick={() => setToast(null)} className="text-gray-400 hover:text-gray-600 mt-0.5 shrink-0">
-              ✕
-            </button>
-          </div>
-        </div>
-      )}
-    </main>
+    </>
   );
 }
